Split ev/core/brain.py into a package by domain
Extracts the Brain class's transcription/OCR/receipt, provider health
checks, LLM-facing tools, one-off ask/plan-day, and multi-provider
fallback chain into mixins under ev/core/brain/, composed in base.py.
_tool_callables and _openai_tools stay together in tools.py since they
are intentional duplicates that must remain synchronized; a new test
(tests/test_brain_tools_sync.py) guards that invariant. Public import
path (ev.core.brain.Brain) is unchanged via re-export.
