+++ ev/core/brain/tools.py
@@ -1,52 +1,17 @@
-"""E.V.'s brain — orchestrates LLM + memory + personality + tools.
+"""LLM-facing tools, in two synchronized formats.
 
-Reusable layer: any interface (Telegram, terminal, web) calls
-`Brain.respond(...)` with text or audio and gets back the answer as text.
-
-Multi-provider strategy (maximize free requests, never go silent):
-  1. GEMINI (primary) — smartest, hears audio natively, saves memory via the
-     google-genai SDK's automatic function calling.
-  2. GROQ (fallback) — Llama 3.3 70B, fast, 30 req/min. When Gemini fails
-     (rate limit, etc.), E.V. keeps talking here, WITH memory (OpenAI-style
-     function calling). Audio is transcribed first via Groq Whisper.
-  3. OPENROUTER (fallback) — plain text, no memory.
-  4. OLLAMA (final safety net) — local model, never rate-limited ("never runs out").
-
-Providers without a configured key are skipped. The brain also does RAG: it
-injects the user's most relevant knowledge-base chunks into the system prompt.
+`_tool_callables` (Gemini-native closures) and `_openai_tools` (OpenAI-schema
+format for Groq/OpenRouter) are intentional duplicates that must stay in sync
+— kept together in this one file, in the same relative order, on purpose.
+See tests/test_brain_tools_sync.py for the guard on this invariant.
 """
 
-# NOTE: intentionally NO `from __future__ import annotations` here. Gemini's
-# automatic function calling introspects the tool functions' annotations at
-# runtime; PEP 563 stringized annotations make it do isinstance(value, "str")
-# -> "isinstance() arg 2 must be a type..." and every tool call fails.
-
-import asyncio
 import logging
-from datetime import datetime, timezone
 
-try:
-    from zoneinfo import ZoneInfo
-except Exception:  # pragma: no cover
-    ZoneInfo = None  # type: ignore
-
-from google import genai
-from google.genai import types
-
-from ..config import Config
-from ..personality import SYSTEM_PROMPT
-from ..providers import documents as documents_mod, embeddings, llm as providers, tools as tools_mod
-from . import knowledge
-from .commands import Commands
-from .memory import Memory
+from ...providers import documents as documents_mod, tools as tools_mod
+from .. import knowledge
 
 log = logging.getLogger("ev.brain")
-
-# Last resort: every provider failed (or no fallback keys configured).
-_ALL_DOWN_MSG = (
-    "Opa, todos os meus cérebros estão no limite agora (o plano grátis tem cota "
-    "por minuto). Me dá uns segundos e tenta de novo, tá?"
-)
 
 # Commands that can't run in the pure logic layer (they send files, drive the
 # Telegram UI, or run a live timer). The LLM's executar_comando queues these and
@@ -58,447 +23,7 @@ _INTERFACE_COMMANDS = frozenset({
 })
 
 
-class Brain:
-    def __init__(self, config: Config, memory: Memory) -> None:
-        self._config = config
-        self._client = genai.Client(api_key=config.gemini_api_key)
-        self._model = config.model
-        self._memory = memory
-        self._commands = Commands(config, memory)  # for the hands-free command tool
-        self._last_provider: str | None = None  # which provider answered last
-        # Documents the LLM asked to create during the current turn. The interface
-        # drains this after respond() and sends each file to the user.
-        self._last_documents: list[dict] = []
-        # Interface-level commands the LLM requested (foco, exportar, status...).
-        # The interface drains these after respond() and runs them with chat context.
-        self._last_actions: list[dict] = []
-        # Tool calls made during the last turn (for the live "terminal" view).
-        self._last_steps: list[dict] = []
-
-    def pop_steps(self) -> list[dict]:
-        """Return and clear the tool-call steps of the last turn (para o terminal).
-        Cada item: {tool, args, result}."""
-        st, self._last_steps = self._last_steps, []
-        return st
-
-    def pop_documents(self) -> list[dict]:
-        """Return and clear the documents generated during the last turn.
-
-        Each item: {bytes, filename, title, content, saved_kb}."""
-        docs, self._last_documents = self._last_documents, []
-        return docs
-
-    def pop_actions(self) -> list[dict]:
-        """Return and clear interface-command intents from the last turn.
-        Each item: {command, args}."""
-        acts, self._last_actions = self._last_actions, []
-        return acts
-
-    def current_model(self) -> str:
-        """Primary Gemini model — a runtime override (via /modelo) wins over .env."""
-        return self._memory.get_setting("model") or self._model
-
-    async def respond(
-        self,
-        user_id: str,
-        *,
-        conv_id: str | None = None,
-        text: str | None = None,
-        audio: bytes | None = None,
-        audio_mime: str | None = None,
-        image: bytes | None = None,
-        image_mime: str | None = None,
-    ) -> str:
-        """Produce E.V.'s answer to a message (text, audio, and/or image).
-
-        `user_id` scopes durable data (facts, tasks, tools) — always the owner.
-        `conv_id` scopes the CONVERSATION thread (defaults to user_id); pass the
-        Telegram chat id so each group/chat keeps its own separate context.
-
-        Runs the blocking SDK calls in a thread so the async event loop is free.
-        """
-        return await asyncio.to_thread(
-            self._respond_sync, user_id, conv_id or user_id,
-            text, audio, audio_mime, image, image_mime,
-        )
-
-    async def ask(self, system: str, prompt: str) -> str | None:
-        """One-off LLM call (no memory/tools) through the provider chain.
-        Used by features like quizzes and weekly insights."""
-        return await asyncio.to_thread(self._ask_sync, system, prompt)
-
-    async def transcribe(self, audio: bytes, mime: str | None) -> str | None:
-        """Transcribe an audio file to text (Groq Whisper). For /transcrever."""
-        return await asyncio.to_thread(self._transcribe, audio, mime)
-
-    async def ocr_image(self, image: bytes, mime: str | None) -> str | None:
-        """Extract text from an image via Gemini vision (OCR)."""
-        return await asyncio.to_thread(self._ocr_sync, image, mime)
-
-    async def plan_day(self, user_id: str) -> str:
-        """Agentic synthesis: an actionable plan from tasks + agenda + weather + loc."""
-        return await asyncio.to_thread(self._plan_day_sync, user_id)
-
-    def _plan_day_sync(self, user_id: str) -> str:
-        cfg = self._config
-        parts = []
-        tasks = self._memory.open_tasks(user_id)
-        if tasks:
-            parts.append("Tarefas abertas:\n" + "\n".join(
-                f"- {t['text']} ({t.get('category', 'geral')})" for t in tasks[:20]))
-        rems = self._memory.open_reminders(user_id)
-        if rems:
-            parts.append("Lembretes:\n" + "\n".join(
-                f"- {r['text']}" + (f" — {r['when_iso'][:16].replace('T', ' ')}"
-                                    if r.get("when_iso") else "") for r in rems[:15]))
-        if cfg.google_authorized():
-            try:
-                parts.append("Agenda:\n" + tools_mod.calendar_upcoming(
-                    cfg, cfg.default_account, 8))
-            except Exception:
-                pass
-        if getattr(cfg, "city", ""):
-            try:
-                parts.append("Clima:\n" + tools_mod.weather(cfg.city))
-            except Exception:
-                pass
-        addr = self._memory.get_setting("loc_addr")
-        if addr:
-            parts.append("Localização atual: " + addr)
-        context = f"(agora: {self._now_str()})\n\n" + (
-            "\n\n".join(parts) or "Sem dados no momento.")
-        system = (
-            "Você é a E.V., assistente pessoal do Ryan. Com base nos dados, monte um "
-            "PLANO curto e acionável para o dia dele, em português do Brasil: priorize "
-            "as tarefas, encaixe-as nos espaços entre os compromissos, avise sobre "
-            "conflitos de horário, clima ou trânsito, e termine com UMA sugestão do que "
-            "fazer AGORA. Use bullets curtos, direto ao ponto. Chame ele de Ryan.")
-        return self._ask_sync(system, context) or "Não consegui montar o plano agora."
-
-    async def describe_image(self, image: bytes, mime: str | None, prompt: str) -> str:
-        """One-off vision description (for live camera / 'what is this'). NOT
-        persisted to any conversation. Returns '' on failure."""
-        return await asyncio.to_thread(self._describe_sync, image, mime, prompt)
-
-    def _describe_sync(self, image: bytes, mime: str | None, prompt: str) -> str:
-        try:
-            resp = self._client.models.generate_content(
-                model=self.current_model(),
-                contents=[
-                    types.Part.from_bytes(data=image, mime_type=mime or "image/jpeg"),
-                    types.Part.from_text(text=prompt),
-                ],
-                config=types.GenerateContentConfig(temperature=0.2),
-            )
-            return (resp.text or "").strip()
-        except Exception as exc:
-            log.warning("describe_image failed (%s)", exc)
-            return ""
-
-    async def extract_receipt(self, image: bytes, mime: str | None) -> dict | None:
-        """Read a receipt/invoice image -> {amount, description, category} or None."""
-        return await asyncio.to_thread(self._extract_receipt_sync, image, mime)
-
-    @staticmethod
-    def _parse_receipt_json(raw: str) -> dict | None:
-        """Parse the vision model's JSON reply into a validated expense dict."""
-        import json
-        import re
-        m = re.search(r"\{.*\}", raw or "", re.S)
-        if not m:
-            return None
-        try:
-            d = json.loads(m.group(0))
-        except (ValueError, TypeError):
-            return None
-        try:
-            amount = round(float(str(d.get("valor", 0)).replace(",", ".").strip()), 2)
-        except (ValueError, TypeError):
-            return None
-        if amount <= 0:
-            return None
-        desc = (str(d.get("descricao") or "").strip() or "compra")[:80]
-        cat = re.sub(r"[^a-zà-ú0-9]", "",
-                     str(d.get("categoria") or "geral").strip().lower()) or "geral"
-        return {"amount": amount, "description": desc, "category": cat}
-
-    def _extract_receipt_sync(self, image: bytes, mime: str | None) -> dict | None:
-        prompt = (
-            "Esta imagem é um comprovante, nota fiscal ou recibo. Extraia o gasto e "
-            "responda APENAS um JSON, sem texto ao redor, com as chaves: "
-            "\"valor\" (número — o TOTAL pago), "
-            "\"descricao\" (estabelecimento ou o que foi comprado, curto), "
-            "\"categoria\" (UMA palavra: mercado, comida, transporte, saude, lazer, "
-            "casa, assinatura, etc). "
-            "Se não for um comprovante ou não achar o total, responda {\"valor\": 0}."
-        )
-        try:
-            resp = self._client.models.generate_content(
-                model=self.current_model(),
-                contents=[
-                    types.Part.from_bytes(data=image, mime_type=mime or "image/jpeg"),
-                    types.Part.from_text(text=prompt),
-                ],
-                config=types.GenerateContentConfig(temperature=0.0),
-            )
-            return self._parse_receipt_json(resp.text or "")
-        except Exception as exc:
-            log.warning("receipt extraction (Gemini vision) failed (%s)", exc)
-            return None
-
-    async def health_check(self) -> list[dict]:
-        """Live-ping each configured provider. Returns [{name, ok, note}]."""
-        return await asyncio.to_thread(self._health_check_sync)
-
-    def _health_check_sync(self) -> list[dict]:
-        cfg = self._config
-        out: list[dict] = []
-        if cfg.gemini_api_key:
-            out.append(self._ping_gemini())
-        if cfg.groq_api_key:
-            out.append(self._ping_openai(
-                "Groq", providers.GROQ_BASE_URL, cfg.groq_api_key, cfg.groq_model))
-        if cfg.openrouter_api_key:
-            out.append(self._ping_openai(
-                "OpenRouter", providers.OPENROUTER_BASE_URL,
-                cfg.openrouter_api_key, cfg.openrouter_model))
-        if cfg.ollama_enabled:
-            out.append(self._ping_openai(
-                "Ollama", cfg.ollama_base_url, "ollama", cfg.ollama_model))
-        if cfg.tavily_api_key:
-            out.append(self._ping_tavily())
-        return out
-
-    @staticmethod
-    def _res(name, ok, note=""):
-        return {"name": name, "ok": bool(ok), "note": note}
-
-    def _ping_gemini(self) -> dict:
-        try:
-            self._client.models.generate_content(
-                model=self.current_model(), contents="ping",
-                config=types.GenerateContentConfig(temperature=0.0),
-            )
-            return self._res("Gemini", True, "respondeu")
-        except Exception as exc:
-            msg = str(exc)
-            rate = "429" in msg or "RESOURCE_EXHAUSTED" in msg
-            # Rate-limit is expected on the free tier — not a real failure.
-            return self._res("Gemini", rate, "cota do dia (normal)" if rate else msg[:70])
-
-    def _ping_openai(self, name, base, key, model) -> dict:
-        try:
-            ans = providers.chat_openai_compat(
-                base_url=base, api_key=key, model=model,
-                system="", messages=[{"role": "user", "content": "ping"}],
-            )
-            return self._res(name, bool(ans), "respondeu" if ans else "sem resposta")
-        except Exception as exc:
-            return self._res(name, False, str(exc)[:70])
-
-    def _ping_tavily(self) -> dict:
-        try:
-            txt = tools_mod.tavily_search("teste", self._config.tavily_api_key, max_results=1)
-            return self._res("Tavily", bool(txt), "respondeu" if txt else "sem resposta")
-        except Exception as exc:
-            return self._res("Tavily", False, str(exc)[:70])
-
-    def _ocr_sync(self, image: bytes, mime: str | None) -> str | None:
-        try:
-            resp = self._client.models.generate_content(
-                model=self.current_model(),
-                contents=[
-                    types.Part.from_bytes(data=image, mime_type=mime or "image/jpeg"),
-                    types.Part.from_text(
-                        text="Extraia TODO o texto visível nesta imagem, exatamente "
-                        "como está, preservando as quebras de linha. Não comente nem "
-                        "resuma. Se não houver texto, responda apenas: (sem texto)"
-                    ),
-                ],
-                config=types.GenerateContentConfig(temperature=0.0),
-            )
-            return (resp.text or "").strip() or None
-        except Exception as exc:
-            log.warning("OCR (Gemini vision) failed (%s)", exc)
-            return None
-
-    def _ask_sync(self, system: str, prompt: str) -> str | None:
-        try:
-            resp = self._client.models.generate_content(
-                model=self.current_model(),
-                contents=prompt,
-                config=types.GenerateContentConfig(
-                    system_instruction=system, temperature=0.5
-                ),
-            )
-            if (resp.text or "").strip():
-                return resp.text.strip()
-        except Exception as exc:
-            log.warning("ask_once Gemini failed (%s)", exc)
-
-        cfg = self._config
-        messages = [{"role": "user", "content": prompt}]
-        chain = []
-        if cfg.groq_api_key:
-            chain.append((providers.GROQ_BASE_URL, cfg.groq_api_key, cfg.groq_model))
-        if cfg.openrouter_api_key:
-            chain.append(
-                (providers.OPENROUTER_BASE_URL, cfg.openrouter_api_key, cfg.openrouter_model)
-            )
-        if cfg.ollama_enabled:
-            chain.append((cfg.ollama_base_url, "ollama", cfg.ollama_model))
-        for base, key, model in chain:
-            try:
-                ans = providers.chat_openai_compat(
-                    base_url=base, api_key=key, model=model,
-                    system=system, messages=messages,
-                )
-                if ans:
-                    return ans
-            except Exception as exc:
-                log.warning("ask_once fallback failed (%s)", exc)
-        return None
-
-    # -----------------------------------------------------------------------
-
-    def _respond_sync(
-        self,
-        user_id: str,
-        conv_id: str,
-        text: str | None,
-        audio: bytes | None,
-        audio_mime: str | None,
-        image: bytes | None = None,
-        image_mime: str | None = None,
-    ) -> str:
-        self._last_documents = []  # fresh per turn; interface drains after respond()
-        self._last_actions = []
-        self._last_steps = []
-        # Semantic recall uses the text query; audio/image-through-Gemini has none yet.
-        system_instruction = self._system_instruction(user_id, text)
-        if text is not None:
-            user_repr = text
-        elif image is not None:
-            user_repr = "[imagem]"
-        else:
-            user_repr = "[mensagem de voz]"
-        answer: str | None = None
-        force = (self._memory.get_setting("force_provider") or "").strip().lower()
-
-        if force == "gemini":
-            # Forced Gemini only — no fallback (so you can actually test it).
-            try:
-                answer = self._gemini(
-                    user_id, conv_id, text, audio, audio_mime, image, image_mime,
-                    system_instruction,
-                )
-            except Exception as exc:
-                log.warning("Forced Gemini failed (%s)", exc)
-        elif force in ("groq", "openrouter", "ollama"):
-            # Forced fallback provider only. Audio -> transcribe; image unsupported.
-            fb_text = text
-            if audio is not None:
-                fb_text = self._transcribe(audio, audio_mime)
-                if fb_text:
-                    user_repr = fb_text
-                    system_instruction = self._system_instruction(user_id, fb_text)
-            if image is not None and not fb_text:
-                return f"O provedor forçado ({force}) não enxerga imagens. Use /provedor gemini ou /provedor auto."
-            if fb_text:
-                answer = self._fallbacks(
-                    user_id, conv_id, fb_text, system_instruction, only=force
-                )
-        else:
-            # Automatic chain: Gemini -> Groq -> OpenRouter -> Ollama.
-            try:
-                answer = self._gemini(
-                    user_id, conv_id, text, audio, audio_mime, image, image_mime,
-                    system_instruction,
-                )
-            except Exception as exc:
-                log.warning("Gemini failed (%s). Trying fallbacks...", exc)
-                fb_text = text
-                if audio is not None:
-                    fb_text = self._transcribe(audio, audio_mime)
-                    if fb_text:
-                        user_repr = fb_text
-                        system_instruction = self._system_instruction(user_id, fb_text)
-                if image is not None and not fb_text:
-                    return "Consegui receber a imagem, mas meu cérebro de visão está no limite agora. Tenta de novo em uns segundos?"
-                if fb_text:
-                    answer = self._fallbacks(user_id, conv_id, fb_text, system_instruction)
-
-        if not answer:
-            if force:
-                return (f"O provedor forçado ({force}) não respondeu agora (pode estar "
-                        "sem cota ou fora do ar). Volta pro automático com /provedor auto.")
-            return _ALL_DOWN_MSG
-
-        # Track which provider answered (for /modelo usage stats).
-        try:
-            self._memory.bump_usage(
-                self._last_provider or "?",
-                datetime.now(timezone.utc).date().isoformat(),
-            )
-        except Exception:
-            pass
-
-        # Persist the turn in this conversation's history (scoped by conv_id).
-        self._memory.add_message(conv_id, "user", user_repr)
-        self._memory.add_message(conv_id, "model", answer)
-        return answer
-
-    # --- system prompt ------------------------------------------------------
-
-    def _system_instruction(self, user_id: str, query: str | None) -> str:
-        system = SYSTEM_PROMPT
-
-        # MODO FOCO — muda o tom das respostas enquanto ativo.
-        if self._memory.get_setting("serious_mode") == "1":
-            system += (
-                "\n\n## MODO FOCO ATIVO\n"
-                "O usuário ativou o modo foco. Responda de forma direta, "
-                "concisa e tática — sem piadas, sem floreio, sem emojis. Vá "
-                "direto ao ponto, tom de operação/missão, frases curtas. "
-                "Continue prestativa e precisa, só que séria e focada."
-            )
-
-        # Current date/time so the model can resolve "tomorrow at 9am" to ISO.
-        system += "\n\n## Data e hora atual\n" + self._now_str()
-
-        # Relevant facts (semantic recall when we have a text query + embeddings).
-        query_vec = self._embed(query) if query else None
-        facts = self._memory.relevant_facts(user_id, query_vec, k=8)
-        if facts:
-            system += "\n\n## O que você já sabe sobre o usuário\n"
-            system += "\n".join(f"- {f}" for f in facts)
-
-        # Knowledge base (RAG): inject the most relevant document chunks.
-        chunks = self._memory.search_knowledge(user_id, query_vec, k=4)
-        if chunks:
-            system += (
-                "\n\n## Trechos relevantes dos documentos do usuário\n"
-                "Use isto para embasar a resposta quando fizer sentido.\n"
-            )
-            for c in chunks:
-                system += f"\n[{c['source']}]\n{c['chunk']}\n"
-        return system
-
-    def _now_str(self) -> str:
-        try:
-            tz = ZoneInfo(self._config.timezone) if ZoneInfo else None
-            now = datetime.now(tz)
-            return (
-                f"Agora é {now.isoformat(timespec='minutes')} "
-                f"(fuso {self._config.timezone}). "
-                "Ao criar lembretes ou eventos, converta para ISO 8601."
-            )
-        except Exception:
-            return "Ao criar lembretes ou eventos, use ISO 8601."
-
-    def _embed(self, text: str) -> list[float] | None:
-        return embeddings.embed(text, self._config)
-
+class ToolsMixin:
     # --- tools (shared by Gemini and Groq) ---------------------------------
 
     def _tool_callables(self, user_id: str) -> dict:
@@ -820,7 +345,7 @@ class Brain:
             Args:
                 nome: nome (ou parte) da playlist.
             """
-            from ..providers import spotify as _sp
+            from ...providers import spotify as _sp
             tok = _sp.access_token(self._memory, self._config)
             if not tok:
                 return "o Spotify não está conectado — conecte na aba Música."
@@ -841,7 +366,7 @@ class Brain:
             Args:
                 busca: o que tocar (faixa e/ou artista).
             """
-            from ..providers import spotify as _sp
+            from ...providers import spotify as _sp
             tok = _sp.access_token(self._memory, self._config)
             if not tok:
                 return "o Spotify não está conectado — conecte na aba Música."
@@ -860,7 +385,7 @@ class Brain:
             Args:
                 acao: 'pausar', 'continuar', 'proxima' ou 'anterior'.
             """
-            from ..providers import spotify as _sp
+            from ...providers import spotify as _sp
             tok = _sp.access_token(self._memory, self._config)
             if not tok:
                 return "o Spotify não está conectado."
@@ -876,7 +401,7 @@ class Brain:
 
         def musica_atual() -> str:
             """O que está tocando agora no Spotify do usuário."""
-            from ..providers import spotify as _sp
+            from ...providers import spotify as _sp
             tok = _sp.access_token(self._memory, self._config)
             if not tok:
                 return "o Spotify não está conectado."
@@ -926,7 +451,7 @@ class Brain:
             import os as _os
             import re as _re
             import json as _json
-            from ..providers import connectors as _cn
+            from ...providers import connectors as _cn
             c = self._memory.get_connector(user_id, nome)
             if not c:
                 av = [x["name"] for x in self._memory.list_connectors(user_id)]
@@ -1354,214 +879,3 @@ class Brain:
                 ),
             ]
         return schemas
-
-    # --- primary provider: Gemini ------------------------------------------
-
-    def _gemini(
-        self,
-        user_id: str,
-        conv_id: str,
-        text: str | None,
-        audio: bytes | None,
-        audio_mime: str | None,
-        image: bytes | None,
-        image_mime: str | None,
-        system_instruction: str,
-    ) -> str:
-        """Call Gemini with memory (function calling). Raises on failure (rate
-        limit, etc.) so the caller falls through to the fallbacks."""
-        tools = list(self._tool_callables(user_id).values())  # data scoped to owner
-        contents = self._build_contents(  # history scoped to this conversation
-            conv_id, text, audio, audio_mime, image, image_mime
-        )
-
-        response = self._client.models.generate_content(
-            model=self.current_model(),
-            contents=contents,
-            config=types.GenerateContentConfig(
-                system_instruction=system_instruction,
-                tools=tools,
-                temperature=0.4,
-            ),
-        )
-        self._last_provider = "gemini"
-        # Log what Gemini's automatic function calling actually did (it is otherwise
-        # opaque) — invaluable for debugging why a CRUD tool didn't run.
-        try:
-            for h in (getattr(response, "automatic_function_calling_history", None) or []):
-                for part in (getattr(h, "parts", None) or []):
-                    fc = getattr(part, "function_call", None)
-                    if fc:
-                        log.info("[gemini-afc] chamou %s args=%s", fc.name,
-                                 dict(fc.args or {}))
-                        self._last_steps.append({"tool": fc.name,
-                                                 "args": dict(fc.args or {})})
-                    fr = getattr(part, "function_response", None)
-                    if fr:
-                        res = str(getattr(fr, "response", ""))[:200]
-                        log.info("[gemini-afc] resultado %s: %s", fr.name, res)
-                        for s in reversed(self._last_steps):   # anexa ao passo correspondente
-                            if s.get("tool") == fr.name and "result" not in s:
-                                s["result"] = res
-                                break
-        except Exception:
-            pass
-        return (response.text or "").strip() or "…"
-
-    def _build_contents(
-        self,
-        conv_id: str,
-        text: str | None,
-        audio: bytes | None,
-        audio_mime: str | None,
-        image: bytes | None = None,
-        image_mime: str | None = None,
-    ) -> list[types.Content]:
-        contents: list[types.Content] = []
-
-        for msg in self._memory.recent_messages(conv_id, limit=20):
-            contents.append(
-                types.Content(
-                    role=msg["role"],
-                    parts=[types.Part.from_text(text=msg["content"])],
-                )
-            )
-
-        new_parts: list[types.Part] = []
-        if audio is not None:
-            new_parts.append(
-                types.Part.from_bytes(data=audio, mime_type=audio_mime or "audio/ogg")
-            )
-            new_parts.append(
-                types.Part.from_text(
-                    text="(mensagem de voz do usuário — responda ao conteúdo dela)"
-                )
-            )
-        if image is not None:
-            new_parts.append(
-                types.Part.from_bytes(data=image, mime_type=image_mime or "image/jpeg")
-            )
-            if text is None:
-                new_parts.append(
-                    types.Part.from_text(
-                        text="(imagem enviada pelo usuário — descreva/analise e ajude)"
-                    )
-                )
-        if text is not None:
-            new_parts.append(types.Part.from_text(text=text))
-
-        contents.append(types.Content(role="user", parts=new_parts))
-        return contents
-
-    # --- fallbacks: Groq -> OpenRouter --------------------------------------
-
-    def _openai_messages(self, conv_id: str, new_text: str) -> list[dict]:
-        # Keep history short on the fallback path: Groq's free tier caps tokens
-        # per minute (~8k), and a long history + tool schemas blows past it (429).
-        msgs: list[dict] = []
-        for m in self._memory.recent_messages(conv_id, limit=8):
-            role = "assistant" if m["role"] == "model" else "user"
-            msgs.append({"role": role, "content": m["content"]})
-        msgs.append({"role": "user", "content": new_text})
-        return msgs
-
-    def _fallbacks(self, user_id: str, conv_id: str, text: str, system: str,
-                   only: str | None = None) -> str | None:
-        messages = self._openai_messages(conv_id, text)  # history per conversation
-        cfg = self._config
-
-        # 1) Groq — WITH memory/tools (function calling): always-available path.
-        if cfg.groq_api_key and only in (None, "groq"):
-            try:
-                answer = providers.chat_with_tools(
-                    base_url=providers.GROQ_BASE_URL,
-                    api_key=cfg.groq_api_key,
-                    model=cfg.groq_model,
-                    system=system,
-                    messages=messages,
-                    tools=self._openai_tools(),
-                    tool_functions=self._tool_callables(user_id),
-                    temperature=0.2,  # lower -> more reliable tool-calling
-                )
-                if answer:
-                    log.info("Answered via Groq (%s) with tools.", cfg.groq_model)
-                    self._last_provider = "groq"
-                    return answer
-            except Exception as exc:
-                log.warning("Groq fallback failed (%s).", exc)
-
-        # 2) OpenRouter — also WITH tools, so CRUD still works when Groq is rate
-        #    limited. Falls back to plain text if the model can't do tool-calling.
-        if cfg.openrouter_api_key and only in (None, "openrouter"):
-            try:
-                answer = providers.chat_with_tools(
-                    base_url=providers.OPENROUTER_BASE_URL,
-                    api_key=cfg.openrouter_api_key,
-                    model=cfg.openrouter_model,
-                    system=system,
-                    messages=messages,
-                    tools=self._openai_tools(),
-                    tool_functions=self._tool_callables(user_id),
-                    temperature=0.2,
-                )
-                if answer:
-                    log.info("Answered via OpenRouter (%s) with tools.", cfg.openrouter_model)
-                    self._last_provider = "openrouter"
-                    return answer
-            except Exception as exc:
-                log.warning("OpenRouter (tools) failed (%s); trying plain text.", exc)
-                try:
-                    answer = providers.chat_openai_compat(
-                        base_url=providers.OPENROUTER_BASE_URL,
-                        api_key=cfg.openrouter_api_key,
-                        model=cfg.openrouter_model,
-                        system=system,
-                        messages=messages,
-                    )
-                    if answer:
-                        log.info("Answered via OpenRouter (%s), plain.", cfg.openrouter_model)
-                        self._last_provider = "openrouter"
-                        return answer
-                except Exception as exc2:
-                    log.warning("OpenRouter plain failed (%s).", exc2)
-
-        # 3) Ollama — local model, never rate-limited (final safety net).
-        if cfg.ollama_enabled and only in (None, "ollama"):
-            try:
-                answer = providers.chat_openai_compat(
-                    base_url=cfg.ollama_base_url,
-                    api_key="ollama",  # Ollama ignores the key
-                    model=cfg.ollama_model,
-                    system=system,
-                    messages=messages,
-                )
-                if answer:
-                    log.info("Answered via Ollama (%s, local).", cfg.ollama_model)
-                    self._last_provider = "ollama"
-                    return answer
-            except Exception as exc:
-                log.warning("Ollama fallback failed (%s).", exc)
-
-        return None
-
-    def _transcribe(self, audio: bytes, audio_mime: str | None) -> str | None:
-        """Transcribe audio via Groq Whisper (for the fallback path)."""
-        if not self._config.groq_api_key:
-            return None
-        # Whisper detects the format from the filename extension — derive it from
-        # the MIME so browser recordings (webm/mp4) and Telegram voice (ogg) work.
-        ext = {
-            "audio/webm": "webm", "audio/ogg": "ogg", "audio/mp4": "mp4",
-            "audio/mpeg": "mp3", "audio/wav": "wav", "audio/x-wav": "wav",
-            "audio/x-m4a": "m4a", "audio/m4a": "m4a", "audio/aac": "m4a",
-        }.get((audio_mime or "").split(";")[0].strip(), "ogg")
-        try:
-            return providers.transcribe_groq(
-                api_key=self._config.groq_api_key,
-                model=self._config.groq_whisper_model,
-                audio=audio,
-                filename=f"audio.{ext}",
-            )
-        except Exception as exc:
-            log.warning("Transcription (Groq Whisper) failed (%s).", exc)
-            return None
